pyproject.toml
[project]
name = "txt-to-video-lora"
version = "1.0.0"
description = "Text to Video generation pipeline using LoRA and AI models"
authors = [
    {name = "Forest", email = "forest@example.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
dependencies = [
    "tqdm",
    "matplotlib", 
    "requests",
    "python-docx",
    "openpyxl",
    "spacy",
    "openai",
    "pillow",
    "moviepy",
    "numpy",
    "pandas",
    "chardet",
    "azure-cognitiveservices-speech",
    "python-dotenv",
    "pydub",
    "pathlib2",
]

[project.urls]
Homepage = "https://github.com/forest/txt-to-video-lora"
Repository = "https://github.com/forest/txt-to-video-lora"
Documentation = "https://github.com/forest/txt-to-video-lora/docs"

[project.scripts]
txt-to-video = "scripts.auto_pipeline:main"
setup-env = "scripts.setup_env:main"
test-llm = "scripts.test_llm:main"
cleanup = "scripts.cleanup:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
